Creates test case for User story birthBeforedeath and added it to the test suite
