+++ src/GEDCOM/AllTests.java
@@ -5,7 +5,7 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({ TestDates.class })
+@SuiteClasses({ TestDates.class,Userstory3Test.class })
 public class AllTests {
 
 }
